+++ spec/src/default_prop_serializer.spec.ts
@@ -1,6 +1,6 @@
 import chai = require('chai');
 
-import { deflate, inflate, Serializable } from './@lib/serialazy';
+import { deserialize, Serializable, serialize } from './@lib/serialazy';
 
 const { expect } = chai;
 
@@ -21,12 +21,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a boolean primitive', () => {
                     const patient = new Patient(true);
-                    const serialized = deflate(patient);
+                    const serialized = serialize(patient);
                     expect(serialized).to.deep.equal({ married: true });
                 });
 
                 it('deserializes to a boolean primitive', () => {
-                    const deserialized = inflate(Patient, { married: false });
+                    const deserialized = deserialize(Patient, { married: false });
                     expect(deserialized instanceof Patient).to.equal(true);
                     expect(deserialized).to.deep.equal({ married: false });
                 });
@@ -37,12 +37,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a boolean primitive', () => {
                     const patient = new Patient(new Boolean(true) as boolean);
-                    const serialized = deflate(patient);
+                    const serialized = serialize(patient);
                     expect(serialized).to.deep.equal({ married: true });
                 });
 
                 it('deserializes to a boolean primitive', () => {
-                    const deserialized = inflate(Patient, { married: new Boolean(false) as boolean });
+                    const deserialized = deserialize(Patient, { married: new Boolean(false) as boolean });
                     expect(deserialized instanceof Patient).to.equal(true);
                     expect(deserialized).to.deep.equal({ married: false });
                 });
@@ -55,11 +55,11 @@ describe('default property serializer', () => {
 
             it('should fail to serialize', () => {
                 const patient = new Patient(new Date() as any);
-                expect(() => deflate(patient)).to.throw('Unable to serialize property "married": Not a boolean');
+                expect(() => serialize(patient)).to.throw('Unable to serialize property "married": Not a boolean');
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Patient, { married: new Date() as any })).to.throw('Unable to deserialize property "married": Not a boolean');
+                expect(() => deserialize(Patient, { married: new Date() as any })).to.throw('Unable to deserialize property "married": Not a boolean');
             });
 
         });
@@ -81,12 +81,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a number primitive', () => {
                     const patient = new Patient(40);
-                    const serialized = deflate(patient);
+                    const serialized = serialize(patient);
                     expect(serialized).to.deep.equal({ age: 40 });
                 });
 
                 it('deserializes to a number primitive', () => {
-                    const deserialized = inflate(Patient, { age: 45 });
+                    const deserialized = deserialize(Patient, { age: 45 });
                     expect(deserialized instanceof Patient).to.equal(true);
                     expect(deserialized).to.deep.equal({ age: 45 });
                 });
@@ -97,12 +97,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a number primitive', () => {
                     const patient = new Patient(new Number(40) as number);
-                    const serialized = deflate(patient);
+                    const serialized = serialize(patient);
                     expect(serialized).to.deep.equal({ age: 40 });
                 });
 
                 it('deserializes to a number primitive', () => {
-                    const deserialized = inflate(Patient, { age: new Number(45) as number });
+                    const deserialized = deserialize(Patient, { age: new Number(45) as number });
                     expect(deserialized instanceof Patient).to.equal(true);
                     expect(deserialized).to.deep.equal({ age: 45 });
                 });
@@ -115,11 +115,11 @@ describe('default property serializer', () => {
 
             it('should fail to serialize', () => {
                 const patient = new Patient(new Date() as any);
-                expect(() => deflate(patient)).to.throw('Unable to serialize property "age": Not a number');
+                expect(() => serialize(patient)).to.throw('Unable to serialize property "age": Not a number');
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Patient, { age: new Date() as any })).to.throw('Unable to deserialize property "age": Not a number');
+                expect(() => deserialize(Patient, { age: new Date() as any })).to.throw('Unable to deserialize property "age": Not a number');
             });
 
         });
@@ -141,12 +141,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a string literal', () => {
                     const greeter = new Greeter('hello');
-                    const serialized = deflate(greeter);
+                    const serialized = serialize(greeter);
                     expect(serialized).to.deep.equal({ message: 'hello' });
                 });
 
                 it('deserializes to a string literal', () => {
-                    const deserialized = inflate(Greeter, { message: 'hi' });
+                    const deserialized = deserialize(Greeter, { message: 'hi' });
                     expect(deserialized instanceof Greeter).to.equal(true);
                     expect(deserialized).to.deep.equal({ message: 'hi' });
                 });
@@ -157,12 +157,12 @@ describe('default property serializer', () => {
 
                 it('serializes to a string literal', () => {
                     const greeter = new Greeter(new String('hello') as string);
-                    const serialized = deflate(greeter);
+                    const serialized = serialize(greeter);
                     expect(serialized).to.deep.equal({ message: 'hello' });
                 });
 
                 it('deserializes to a string literal', () => {
-                    const deserialized = inflate(Greeter, { message: new String('hello') as string });
+                    const deserialized = deserialize(Greeter, { message: new String('hello') as string });
                     expect(deserialized instanceof Greeter).to.equal(true);
                     expect(deserialized).to.deep.equal({ message: 'hello' });
                 });
@@ -175,11 +175,11 @@ describe('default property serializer', () => {
 
             it('should fail to serialize', () => {
                 const greeter = new Greeter(new Date() as any);
-                expect(() => deflate(greeter)).to.throw('Unable to serialize property "message": Not a string');
+                expect(() => serialize(greeter)).to.throw('Unable to serialize property "message": Not a string');
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Greeter, { message: new Date() as any })).to.throw('Unable to deserialize property "message": Not a string');
+                expect(() => deserialize(Greeter, { message: new Date() as any })).to.throw('Unable to deserialize property "message": Not a string');
             });
 
         });
@@ -214,12 +214,12 @@ describe('default property serializer', () => {
             const book = new Book('The Story of the Sealed Room', new Author('Arthur Conan Doyle'));
 
             it('serializes to JSON-compatible object', () => {
-                const serialized = deflate(book);
+                const serialized = serialize(book);
                 expect(serialized).to.deep.equal(bookObj);
             });
 
             it('deserializes from JSON-compatible object', () => {
-                const deserialized = inflate(Book, bookObj);
+                const deserialized = deserialize(Book, bookObj);
                 expect(deserialized instanceof Book).to.equal(true);
                 expect(deserialized).to.deep.equal(book);
             });
@@ -247,11 +247,11 @@ describe('default property serializer', () => {
             const book = new Book('The Story of the Sealed Room', new Author('Arthur Conan Doyle'));
 
             it('should fail to serialize', () => {
-                expect(() => deflate(book)).to.throw('No serializer found for type: "Author"');
+                expect(() => serialize(book)).to.throw('No serializer found for type: "Author"');
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Book, bookObj)).to.throw('No serializer found for type: "Author"');
+                expect(() => deserialize(Book, bookObj)).to.throw('No serializer found for type: "Author"');
             });
 
         });