+++ spec/src/facade_functions.spec.ts
@@ -1,6 +1,6 @@
 import chai = require('chai');
 
-import { deflate, inflate, Serializable } from './@lib/serialazy';
+import { deserialize, Serializable, serialize } from './@lib/serialazy';
 
 const { expect } = chai;
 
@@ -14,73 +14,73 @@ describe('facade function', () => {
         public id: string;
     }
 
-    describe('deflate', () => {
+    describe('serialize', () => {
 
         it('is able to serialize null/undefined', () => {
-            expect(deflate(null)).to.equal(null);
-            expect(deflate(undefined)).to.equal(undefined);
+            expect(serialize(null)).to.equal(null);
+            expect(serialize(undefined)).to.equal(undefined);
         });
 
         it('is able to serialize primitives', () => {
 
-            expect(deflate(true)).to.be.a('boolean');
-            expect(deflate(new Boolean(true))).to.be.a('boolean');
-            expect(deflate(true)).to.equal(true);
-            expect(deflate(new Boolean(true))).to.equal(true);
+            expect(serialize(true)).to.be.a('boolean');
+            expect(serialize(new Boolean(true))).to.be.a('boolean');
+            expect(serialize(true)).to.equal(true);
+            expect(serialize(new Boolean(true))).to.equal(true);
 
-            expect(deflate(12)).to.be.a('number');
-            expect(deflate(new Number(12))).to.be.a('number');
-            expect(deflate(12)).to.equal(12);
-            expect(deflate(new Number(12))).to.equal(12);
+            expect(serialize(12)).to.be.a('number');
+            expect(serialize(new Number(12))).to.be.a('number');
+            expect(serialize(12)).to.equal(12);
+            expect(serialize(new Number(12))).to.equal(12);
 
-            expect(deflate('hello')).to.be.a('string');
-            expect(deflate(new String('hello'))).to.be.a('string');
-            expect(deflate('hello')).to.equal('hello');
-            expect(deflate(new String('hello'))).to.equal('hello');
+            expect(serialize('hello')).to.be.a('string');
+            expect(serialize(new String('hello'))).to.be.a('string');
+            expect(serialize('hello')).to.equal('hello');
+            expect(serialize(new String('hello'))).to.equal('hello');
 
         });
 
         it('is able to serialize non-primitives which are serializable', () => {
             const foo = Object.assign(new Foo(), { id: 'foo' });
-            expect(deflate(foo)).to.deep.equal({ id: 'foo' });
+            expect(serialize(foo)).to.deep.equal({ id: 'foo' });
         });
 
         it('should fail to serialize non-primitives which are not serializable', () => {
             const bar = Object.assign(new Bar(), { id: 'bar' });
-            expect(() => deflate(bar)).to.throw("Value is not serializable");
+            expect(() => serialize(bar)).to.throw("Value is not serializable");
         });
 
     });
 
-    describe('inflate', () => {
+    describe('deserialize', () => {
 
         it('is able to deserialize primitives', () => {
 
-            expect(inflate(Boolean, true)).to.be.a('boolean');
-            expect(inflate(Boolean, new Boolean(true) as boolean)).to.be.a('boolean');
-            expect(inflate(Boolean, true)).to.equal(true);
-            expect(inflate(Boolean, new Boolean(true) as boolean)).to.equal(true);
+            expect(deserialize(Boolean, true)).to.be.a('boolean');
+            expect(deserialize(Boolean, new Boolean(true) as boolean)).to.be.a('boolean');
+            expect(deserialize(Boolean, true)).to.equal(true);
+            expect(deserialize(Boolean, new Boolean(true) as boolean)).to.equal(true);
 
-            expect(inflate(Number, 12)).to.be.a('number');
-            expect(inflate(Number, new Number(12) as number)).to.be.a('number');
-            expect(inflate(Number, 12)).to.equal(12);
-            expect(inflate(Number, new Number(12) as number)).to.equal(12);
+            expect(deserialize(Number, 12)).to.be.a('number');
+            expect(deserialize(Number, new Number(12) as number)).to.be.a('number');
+            expect(deserialize(Number, 12)).to.equal(12);
+            expect(deserialize(Number, new Number(12) as number)).to.equal(12);
 
-            expect(inflate(String, 'hello')).to.be.a('string');
-            expect(inflate(String, new String('hello') as string)).to.be.a('string');
-            expect(inflate(String, 'hello')).to.equal('hello');
-            expect(inflate(String, new String('hello') as string)).to.equal('hello');
+            expect(deserialize(String, 'hello')).to.be.a('string');
+            expect(deserialize(String, new String('hello') as string)).to.be.a('string');
+            expect(deserialize(String, 'hello')).to.equal('hello');
+            expect(deserialize(String, new String('hello') as string)).to.equal('hello');
 
         });
 
         it('is able to deserialize non-primitives which are serializable', () => {
-            const foo = inflate(Foo, { id: 'foo' });
+            const foo = deserialize(Foo, { id: 'foo' });
             expect(foo).to.be.instanceOf(Foo);
             expect(foo).to.deep.equal({ id: 'foo' });
         });
 
         it('should fail to deserialize non-primitives which are not serializable', () => {
-            expect(() => inflate(Bar, { id: 'bar' })).to.throw('Type is not serializable: Bar');
+            expect(() => deserialize(Bar, { id: 'bar' })).to.throw('Type is not serializable: Bar');
         });
 
     });