+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* [BREAKING] Renamed `deflate` to `serialize` (alias `serialize.toJSON`), `inflate` to `deserialize` (alias `deserialize.fromJSON`)
 * [BREAKING] Renamed property/type decorators.
   `@Serialize()` to `@Serializable.JSON.Prop()` (alias `@Serializable.Prop()`)
   `@Serialize.Type()` to `@Serializable.JSON.Type()` (alias `@Serializable.Type()`)